test(extraction): move TCP-server cross-loop regression tests to integration (closes #329)
